// src/app/models/studentAssignment.ts
export class StudentAssignment {
    id: number;
    branchId:number;
    courseId: number;
    batchId: number;  
    studentId: number;
    courseFeeId: number;
    receiptId: number;
    studentName: string;
    courseName:  string;
    batchName:  string;
    feeStructureName :  string;
    createdDate: string;
    updatedDate: string;
    createdBy: number;
    updatedBy: number;
  }
  